% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brok_scrape.R
\name{brok_scrape}
\alias{brok_scrape}
\title{Brok Scrape}
\usage{
brok_scrape(urls)
}
\arguments{
\item{urls}{a vector of target URLs as strings}
}
\value{
scrapes: a tibble containing the original URL and the simplified text
}
\description{
Scrapes an HTML site with rvest and returns a tibble containing the original URL and the scraped and simplified text
}
\examples{

brok_scrape(c('https://realpython.github.io/fake-jobs/jobs/energy-engineer-1.html', 'https://realpython.github.io/fake-jobs/jobs/materials-engineer-24.html'))

                           "url                                                                        text
https://realpython.github.io/fake-jobs/jobs/energy-engineer-1.html       | Fake Python Fake Jobs for Your Web Scraping Journey ...
https://realpython.github.io/fake-jobs/jobs/materials-engineer-24.html   | Fake Python Fake Jobs for Your Web Scraping Journey ..."
}
