% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_id.R
\name{create_id}
\alias{create_id}
\title{Create ID}
\usage{
create_id(urls)
}
\arguments{
\item{urls}{A character vector of urls.}
}
\value{
A character vector of unique identifiers.
}
\description{
Convert a character vector of provided urls into a character vector of unique identifiers for use in downstream functions.
}
\examples{
urls <- c('https://www.reddit.com/r/nba/', 'https://www.reddit.com/r/nfl/', 'https://vancouver.craigslist.org/search/apa', 'https://www.kaggle.com/search?q=nba')
create_id(urls)
}
