% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_id.R
\name{create_id}
\alias{create_id}
\title{Create ID}
\usage{
create_id(urls)
}
\arguments{
\item{urls}{A vector of urls as strings.}
}
\value{
A vector of unique identifiers as strings.
}
\description{
Convert a vector of provided urls into a vector of unique identifiers for use in downstream functions.
}
\examples{
urls <- c('https://www.reddit.com/r/nba/', 'https://www.reddit.com/r/nfl/', 'https://vancouver.craigslist.org/search/apa', 'https://www.kijiji.ca/b-real-estate/richmond-bc/c34l1700288')
create_id(urls)
[1] 'reddit1'  'reddit2'  'craigslist1'  'kijiji1'
}
