% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duster.R
\name{duster}
\alias{duster}
\title{Duster}
\usage{
duster(urls)
}
\arguments{
\item{urls}{: vector of target urls as strings}
}
\value{
df : A tibble with the webpage identifiers as a index, the raw url, and the raw text from the webpage with extra line breaks removed
}
\description{
Prepares a pandas dataframe by webscraping raw text from a list of urls ready to be input into a machine learning model
}
\examples{
duster(['https://www.cnn.com/world', 'https://www.foxnews.com/world', 'https://www.cbc.ca/news/world'])
            url                                           raw_text
id
cnn1          https://www.cnn.com/world   World news - breaking news, video, headlines ...
foxnews1  https://www.foxnews.com/world  World | Fox NewsFox News   U.S.PoliticsMediaOp...
cbc1      https://www.cbc.ca/news/world  World - CBC NewsContentSkip to Main ContentAcc...
}
